+++ DriverLevelTool/viewer/renderheightmap.h
@@ -1,0 +1,9 @@
+#ifndef RENDERHEIGHTMAP_H
+#define RENDERHEIGHTMAP_H
+
+#include "math/Vector.h"
+struct VECTOR_NOPAD;
+
+void DebugDrawDriver2HeightmapCell(const VECTOR_NOPAD& cellPos, const ColorRGBA& color);
+
+#endif // RENDERHEIGHTMAP_H
